tests: cleanup input enum tests
remove dead code
add enum prefix to 'duplicated' structs
separated enum tests from non-enum
