+++ tests/test_synctest_session.rs
@@ -1,5 +1,5 @@
-mod enum_input;
 mod stubs;
+mod stubs_enum;
 
 use ggrs::{GGRSError, GGRSRequest, SessionBuilder};
 use stubs::{StubConfig, StubInput};
@@ -100,26 +100,4 @@ fn test_advance_frames_with_random_checksums() {
         stub.handle_requests(requests);
         assert_eq!(stub.gs.frame, i as i32 + 1);
     }
-}
-
-#[test]
-fn test_enum_advance_frames_with_delayed_input() -> Result<(), GGRSError> {
-    let check_distance = 7;
-    let mut stub = enum_input::GameStub::new();
-    let mut sess = SessionBuilder::new()
-        .with_check_distance(check_distance)
-        .with_input_delay(2)
-        .start_synctest_session()?;
-
-    let inputs = [enum_input::EnumInput::Val1, enum_input::EnumInput::Val2];
-    for i in 0..200 {
-        let input = inputs[i % inputs.len()];
-        sess.add_local_input(0, input)?;
-        sess.add_local_input(1, input)?;
-        let requests = sess.advance_frame()?;
-        stub.handle_requests(requests);
-        assert_eq!(stub.gs.frame, i as i32 + 1); // frame should have advanced
-    }
-
-    Ok(())
 }
